+++ common/src/main/java/net/skill_tree_rpgs/skills/NodeTypes.java
@@ -415,7 +415,10 @@ public class NodeTypes {
     public static final Entry ARCHER_TIER_4_SPELL_1_ROOT = add(modifierSpell(ArcherSkills.archer_tier_4_spell_1_root).require(ARCHERS));
     public static final Entry ARCHER_TIER_4_SPELL_2_ROOT = add(modifierSpell(ArcherSkills.archer_tier_4_spell_2_root).require(ARCHERS));
     public static final Entry ARCHER_TIER_4_SPELL_2_MODIFIER_1 = add(modifierSpell(ArcherSkills.archer_tier_4_spell_2_modifier_1).require(ARCHERS));
-    public static final Entry ARCHER_TIER_4_SPELL_2_MODIFIER_2 = add(modifierSpell(ArcherSkills.archer_tier_4_spell_2_modifier_2).require(ARCHERS));
+    public static final Entry ARCHER_TIER_4_SPELL_2_MODIFIER_2 = add(passiveSpell(ArcherSkills.archer_tier_4_spell_2_modifier_2)
+            .withIcon(Icon.spell(Identifier.of("archers", "rain_of_arrows")))
+            .require(ARCHERS)
+    );
 
     public static final Entry ARCHER_TIER_1_PASSIVE_1 = add(passiveSpell(ArcherSkills.archer_tier_1_passive_1).require(ARCHERS));
     public static final Entry ARCHER_TIER_1_PASSIVE_2 = add(passiveSpell(ArcherSkills.archer_tier_1_passive_2).require(ARCHERS));
@@ -789,8 +792,8 @@ public class NodeTypes {
     public static final Entry WARRIOR_TIER_3_SPELL_1_ROOT = add(modifierSpell(WarriorSkills.warrior_tier_3_spell_1_root).require(ROGUES));
     public static final Entry WARRIOR_TIER_4_SPELL_1_ROOT = add(modifierSpell(WarriorSkills.warrior_tier_4_spell_1_root).require(ROGUES));
     public static final Entry WARRIOR_TIER_2_SPELL_2_ROOT = add(modifierSpell(WarriorSkills.warrior_tier_2_spell_2_root).require(ROGUES));
-    public static final Entry WARRIOR_TIER_2_SPELL_2_MODIFIER_1 = add(placeholderNode("warrior_tier_2_spell_2_modifier_1", ROGUES));
-    public static final Entry WARRIOR_TIER_2_SPELL_2_MODIFIER_2 = add(placeholderNode("warrior_tier_2_spell_2_modifier_2", ROGUES));
+    public static final Entry WARRIOR_TIER_2_SPELL_2_MODIFIER_1 = add(modifierSpell(WarriorSkills.warrior_tier_2_spell_2_modifier_1).require(ROGUES));
+    public static final Entry WARRIOR_TIER_2_SPELL_2_MODIFIER_2 = add(modifierSpell(WarriorSkills.warrior_tier_2_spell_2_modifier_2).require(ROGUES));
     public static final Entry WARRIOR_TIER_3_SPELL_2_ROOT = add(modifierSpell(WarriorSkills.warrior_tier_3_spell_2_root).require(ROGUES));
     public static final Entry WARRIOR_TIER_3_SPELL_2_MODIFIER_1 = add(passiveSpell(WarriorSkills.warrior_tier_3_spell_2_modifier_1)
             .withIcon(Icon.spell(Identifier.of("rogues", "shout")))
@@ -800,8 +803,11 @@ public class NodeTypes {
             .require(ROGUES)
     );
     public static final Entry WARRIOR_TIER_4_SPELL_2_ROOT = add(modifierSpell(WarriorSkills.warrior_tier_4_spell_2_root).require(ROGUES));
-    public static final Entry WARRIOR_TIER_4_SPELL_2_MODIFIER_1 = add(placeholderNode("warrior_tier_4_spell_2_modifier_1", ROGUES));
-    public static final Entry WARRIOR_TIER_4_SPELL_2_MODIFIER_2 = add(placeholderNode("warrior_tier_4_spell_2_modifier_2", ROGUES));
+    public static final Entry WARRIOR_TIER_4_SPELL_2_MODIFIER_1 = add(modifierSpell(WarriorSkills.warrior_tier_4_spell_2_modifier_1).require(ROGUES));
+    public static final Entry WARRIOR_TIER_4_SPELL_2_MODIFIER_2 = add(passiveSpell(WarriorSkills.warrior_tier_4_spell_2_modifier_2)
+            .withIcon(Icon.spell(Identifier.of("rogues", "last_stand")))
+            .require(ROGUES)
+    );
 
     public static final Entry FIREBALL = add(
             Entry.spell("fireball",